Registers
Register module seems to be fairly stable at this point. Types and
factory functions with preconditions coded to minimize code errors
down the line.

Commit at this time because it's time to break the "Register" concept
into smaller modules. Time to hide the type definitions at a lower
level to prevent creation of invalid type instances.